+++ src/main/java/com/example/backendtourservice/repository/tour/TourRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
 public interface TourRepository extends JpaRepository<TourEntity, Long> {
-    List<TourEntity> findByAreaCodeAndSigunguCodeAndContentTypeId(Integer areaCode, Integer sigunguCode, Long contentTypeId);
+    List<TourEntity> findByAreaCodeAndSigunguCodeAndContentTypeId(Integer areaCode, Integer sigunguCode, Integer contentTypeId);
 
     @Query(value = "SELECT * FROM tour.tour_spot WHERE content_type_id = ? AND ST_DWithin("
             + "(ST_MakePoint(CAST(location_x AS double precision), CAST(location_y AS double precision))), "
